% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_to_df.R
\name{df_to_list}
\alias{df_to_list}
\title{Converts a name-value data frame to a named list}
\usage{
df_to_list(df_for_list)
}
\arguments{
\item{df_for_list}{The data frame to be converted to a list}
}
\description{
This function converts a data frame back to a list. It is the reverse
operation to \link{list_to_df}.
}
\details{
In a data frame with more than two columns, heuristics are applied to detect
the name and value column.
}

