\docType{data}
\name{fortytwo}
\alias{fortytwo}
\title{The Answer to the Ultimate Question of Life, the Universe, and
  Everything}
\format{A data frame with 1 row}
\source{
  The Hitchhiker's Guide to the Galaxy
}
\usage{
  fortytwo
}
\description{
  This has been calculated by an enormous supercomputer
  over a period of 7.5 million years. Unfortunately no one
  knows what the question is.
}
\author{
  Douglas Adams
}
\keyword{datasets}

