% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{kimisc-deprecated}
\alias{kimisc-deprecated}
\title{Deprecated functions}
\description{
The "See also" section contains the deprecated functions in this package.
}
\seealso{
Other deprecated functions: 
\code{\link{coalesce.na-deprecated}},
\code{\link{df_to_list-deprecated}},
\code{\link{hms.to.seconds-deprecated}},
\code{\link{list_to_df-deprecated}},
\code{\link{nc-deprecated}},
\code{\link{nlist-deprecated}},
\code{\link{ofactor-deprecated}},
\code{\link{sample.rows-deprecated}},
\code{\link{seconds.to.hms-deprecated}},
\code{\link{thisfile-deprecated}},
\code{\link{tll-deprecated}}
}
\concept{deprecated functions}
