% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimisc-package.R
\docType{package}
\name{kimisc-package}
\alias{kimisc}
\alias{kimisc-package}
\title{kimisc: Kirill's Miscellaneous Functions}
\description{
A collection of useful functions not found anywhere else,
mainly for programming: Pretty intervals, generalized lagged
differences, checking containment in an interval, creating a
factor where the levels maintain the order of appearance,
sampling rows from a data frame, converting seconds from
midnight to and from H:M:S format, choosing the first non-NA
value, transposing lists of lists, returning the name of the
file currently sourced, smart named lists and vectors, and an
alternative interface to assign().
}
\seealso{
Useful links:
\itemize{
  \item \url{http://krlmlr.github.io/kimisc}
  \item Report bugs at \url{https://github.com/krlmlr/kimisc/issues}
}

}
\author{
Kirill Müller
}
