% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_df.R
\name{list_to_df-deprecated}
\alias{list_to_df-deprecated}
\alias{list_to_df}
\title{Converts a list to a name-value data frame}
\usage{
list_to_df(list_for_df)
}
\arguments{
\item{list_for_df}{The object to be converted to a data frame}
}
\description{
This function coerces its input to a list and returns a data frame with as many rows as there are list items in the input,
and two columns (one for the names, one for the values).
If the list is not named, the natural sequence will be used as item names.
}
\seealso{
Other deprecated functions:
\code{\link{coalesce.na-deprecated}},
\code{\link{df_to_list-deprecated}},
\code{\link{hms.to.seconds-deprecated}},
\code{\link{kimisc-deprecated}},
\code{\link{nc-deprecated}},
\code{\link{nlist-deprecated}},
\code{\link{ofactor-deprecated}},
\code{\link{sample.rows-deprecated}},
\code{\link{seconds.to.hms-deprecated}},
\code{\link{thisfile-deprecated}},
\code{\link{tll-deprecated}},
\code{\link{vswitch-deprecated}}
}
\concept{deprecated functions}
\keyword{internal}
