% Generated by roxygen2 (4.0.0): do not edit by hand
\name{nc}
\alias{nc}
\title{Smart named vector}
\usage{
nc(...)
}
\arguments{
\item{...}{Vector elements, possibly named}
}
\value{
A named vector.
}
\description{
This function is a wrapper around \code{\link{c}} that
  assigns names to unnamed arguments based on the unevaluated expression used
  in the call.
}
\examples{
a <- 1; b <- 2; c <- 3
nc(a, b, d=c)
nc(mean(c(a, b, c)))
}
\author{
Hadley Wickham
}
\references{
\url{http://stackoverflow.com/a/5043280/946850},
  \url{http://tolstoy.newcastle.edu.au/R/e9/help/10/03/8392.html}
}
\seealso{
\link{c}, \link{nlist}
}

