% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlist.R
\name{nlist}
\alias{nlist}
\title{Smart named list}
\usage{
nlist(...)
}
\arguments{
\item{...}{List items, possibly named}
}
\value{
A named list.
}
\description{
This function is a wrapper around \code{\link[=list]{list()}} that
assigns names to unnamed arguments based on the unevaluated expression used
in the call.
}
\examples{
a <- 1; b <- 2; c <- 3
nlist(a, b, d=c)
nlist(mean(c(a, b, c)))
}
\references{
http://stackoverflow.com/a/5043280/946850,
http://tolstoy.newcastle.edu.au/R/e9/help/10/03/8392.html
}
\seealso{
\link{list}
}
\author{
Hadley Wickham
}
