% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tll}
\alias{tll}
\title{Transposes a list of lists}
\usage{
tll(l)
}
\arguments{
\item{l}{List of lists, possibly named.}
}
\value{
A list of lists corresponding to a transposition of the argument.
}
\description{
The argument is assumed to be a list of \eqn{n} (named) lists
  with length \eqn{m} each.  It is converted to a (named) list of \eqn{m}
  elements with length \eqn{n} each.
}
\examples{
tll(list(list(1, 2), list(3, 4)))
tll(list(list(a=1, b=2), list(a=3, b=4)))
tll(list(x=list(a=1, b=2), y=list(a=3, b=4)))
}
\seealso{
\link[base]{t}
}

