\docType{package}
\name{kimisc-package}
\alias{kimisc}
\alias{kimisc-package}
\title{My first attempt to create an R package}
\description{
  This package is intended to be a collection of more or
  less useful functions not found in default R or anywhere
  else.
}
\details{
  Version: 0.0.19

  Date: 2013-09-12
}
\examples{
data(fortytwo)
}
\author{
  Kirill Müller
}
\keyword{package}

