% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{kimisc-package}
\alias{kimisc}
\alias{kimisc-package}
\title{Kirill's miscellaneous functions}
\description{
This package is intended to be a collection of more or
  less useful functions not found anywhere else.
}
\details{
\tabular{ll}{
  Package: \tab kimisc\cr
  Type: \tab Package\cr
  Version: \tab 2.0\cr
  Date: \tab 2014-03-20\cr
  License: \tab GPL v3\cr
}
}
\author{
Kirill Müller
}
\keyword{package}

