% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimisc-package.R
\docType{package}
\name{kimisc-package}
\alias{kimisc}
\alias{kimisc-package}
\title{kimisc: Kirill's Miscellaneous Functions}
\description{
A collection of useful functions not found anywhere else, mainly for programming: Pretty intervals, generalized lagged differences, checking containment in an interval, and an alternative interface to assign().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://krlmlr.github.io/kimisc/}
  \item \url{https://github.com/krlmlr/kimisc}
  \item Report bugs at \url{https://github.com/krlmlr/kimisc/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com}

}
\keyword{internal}
