\docType{package}
\name{kimisc-package}
\alias{kimisc}
\alias{kimisc-package}
\title{My first attempt to create an R package}
\description{
  This package is intended to be a collection of more or
  less useful functions not found in default R or anywhere
  else.
}
\details{
  \tabular{ll}{ Package: \tab kimisc\cr Type: \tab
  Package\cr Version: \tab 0.1.3\cr Date: \tab
  2013-09-13\cr License: \tab GPL v3\cr }
}
\examples{
data(fortytwo)
}
\author{
  Kirill Müller
}
\keyword{package}

