% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofactor.R
\name{ofactor-deprecated}
\alias{ofactor-deprecated}
\alias{ofactor}
\title{Order-preserving factors}
\usage{
ofactor(x = character(), ...)
}
\arguments{
\item{x}{A vector of data, usually taking a small number of distinct values.}

\item{...}{Other arguments passed on to \code{\link[base:factor]{base::factor()}}.}
}
\value{
A factor. See \code{\link[base:factor]{base::factor()}} for details.
}
\description{
The function \code{ofactor()} is a convenience wrapper for
\code{\link[=factor]{factor()}} that orders the levels as they appear in the data if the
\code{levels} argument is not specified.
}
\examples{
ofactor(3:1)
ofactor(9:12, exclude=11)
identical(ofactor(3:1, levels=1:3), factor(3:1))
}
\seealso{
Other deprecated functions: \code{\link{coalesce.na-deprecated}},
  \code{\link{df_to_list-deprecated}},
  \code{\link{hms.to.seconds-deprecated}},
  \code{\link{kimisc-deprecated}},
  \code{\link{list_to_df-deprecated}},
  \code{\link{nc-deprecated}},
  \code{\link{nlist-deprecated}},
  \code{\link{sample.rows-deprecated}},
  \code{\link{seconds.to.hms-deprecated}},
  \code{\link{thisfile-deprecated}},
  \code{\link{tll-deprecated}},
  \code{\link{vswitch-deprecated}}
}
