% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ofactor}
\alias{ofactor}
\title{Order-preserving factors}
\usage{
ofactor(x = character(), ...)
}
\arguments{
\item{x}{A vector of data, usually taking a small number of distinct values.}

\item{...}{Other arguments passed on to \code{\link[base]{factor}}.}
}
\value{
A factor. See \code{\link[base]{factor}} for details.
}
\description{
The function \code{ofactor} is a convenience wrapper for
  \code{factor} that orders the levels as they appear in the data if the
  \code{levels} argument is not specified.
}
\examples{
ofactor(3:1)
ofactor(9:12, exclude=11)
identical(ofactor(3:1, levels=1:3), factor(3:1))
}

